+++ backend/backend/server/blueprints/users.py
@@ -2,7 +2,6 @@ import jwt
 import secrets
 import hashlib
 import datetime
-from functools import wraps
 from flask import Blueprint, request, make_response, Response, current_app
 from flask_restful import Api, Resource, abort
 from flask_cors import cross_origin
@@ -10,46 +9,11 @@ from flask_cors import cross_origin
 
 from backend.models import *
 from backend.db import db
-from backend.server.logging import logger
 
 users_blueprint = Blueprint("users", __name__)
 users_api = Api(users_blueprint)
 
 
-def jwt_authenticated(f: typing.Callable) -> typing.Callable:
-    """
-    Decorator to check if a user is authenticated using JWT by checking the Authorization header.
-    The request object must be reachable.
-    """
-
-    @wraps(f)
-    def decorated(*args, **kwargs) -> typing.Callable:
-        token = request.headers.get("Authorization")
-        if not token:
-            abort(
-                401,
-                message="No token provided",
-                headers={"WWW-Authenticate": "Basic realm='Valid login required'"},
-            )
-        try:
-            payload = jwt.decode(
-                token.split(" ")[1],
-                key=current_app.config["ACCESS_KEY_SECRET"],
-                algorithms=["HS256"],
-            )
-            user_id = payload["user"]
-            # TODO: Implement a check against cookie for fingerprint once https is working
-            # if payload["fingerprint"] != request.cookies.get("additional_token"):
-            #     abort(401, message="Invalid token")
-        except jwt.ExpiredSignatureError as ese:
-            logger.error(f"Token has expired: {ese}")
-            abort(401, message="Token has expired, re-login is required.")
-        except jwt.InvalidTokenError as ite:
-            logger.error(f"Invalid token: {ite}")
-            abort(401, message="Invalid token")
-        return f(*args, **kwargs, user_id=user_id)
-
-    return decorated
 
 
 class UserRegistration(Resource):
